resource/参考代码/stone_detection/requirements.txt
absl-py                     1.2.0
aiohttp                     3.8.5
aiosignal                   1.3.1
albumentations              1.3.0
altgraph                    0.17.3
antlr4-python3-runtime      4.9.3
async-timeout               4.0.3
asynctest                   0.13.0
attrs                       23.1.0
av                          10.0.0
backcall                    0.2.0
cachetools                  5.2.0
certifi                     2022.6.15
cffi                        1.15.1
chardet                     5.1.0
charset-normalizer          2.1.1
click                       8.1.3
colorama                    0.4.5
coloredlogs                 15.0.1
cycler                      0.11.0
Cython                      0.29.32
decorator                   5.1.1
docstring-parser            0.15
efficientnet-pytorch        0.7.1
einops                      0.6.0
et-xmlfile                  1.1.0
exceptiongroup              1.1.3
filelock                    3.9.0
flatbuffers                 23.5.26
fonttools                   4.37.4
FrEIA                       0.2
frozenlist                  1.3.3
fsspec                      2023.1.0
gitdb                       4.0.10
GitPython                   3.1.30
google-auth                 2.11.0
google-auth-oauthlib        0.4.6
grpcio                      1.48.1
h11                         0.14.0
huggingface-hub             0.12.1
humanfriendly               10.0
idna                        3.3
imageio                     2.25.1
imgaug                      0.4.0
imgviz                      1.7.2
importlib-metadata          4.12.0
importlib-resources         5.12.0
imutils                     0.5.4
install                     1.3.5
ipython                     7.34.0
jedi                        0.18.2
Jinja2                      3.1.2
joblib                      1.2.0
jsonargparse                4.23.1
kiwisolver                  1.4.4
kornia                      0.6.9
labelme                     5.2.0.post4
lightning-utilities         0.9.0
lmdb                        1.4.1
lxml                        4.9.1
Markdown                    3.4.1
MarkupSafe                  2.1.1
matplotlib                  3.5.3
matplotlib-inline           0.1.6
mc-bin-client               1.0.1
modbus-tk                   1.1.3
mpmath                      1.3.0
multidict                   6.0.4
munch                       2.5.0
natsort                     8.3.1
networkx                    2.6.3
nn                          0.1.1
numpy                       1.21.6
oauthlib                    3.2.0
omegaconf                   2.3.0
onnx                        1.13.0
onnxruntime                 1.14.1
opencv-contrib-python       4.8.0.76
opencv-python               4.8.0.76
opencv-python-headless      4.2.0.34
openpyxl                    3.1.2
outcome                     1.2.0
packaging                   21.3
pandas                      1.1.5
parso                       0.8.3
pefile                      2023.2.7
pickleshare                 0.7.5
Pillow                      9.2.0
pip                         23.0.1
pretrainedmodels            0.7.4
prompt-toolkit              3.0.36
protobuf                    3.20.3
psutil                      5.9.4
pyasn1                      0.4.8
pyasn1-modules              0.2.8
pycocotools-windows         2.0.0.2
pycparser                   2.21
Pygments                    2.14.0
pyinstaller                 5.11.0
pyinstaller-hooks-contrib   2023.3
pymodbus                    2.5.3
PyMySQL                     1.1.0
pyparsing                   3.0.9
PyQt5                       5.15.9
pyqt5-plugins               5.15.9.2.3
PyQt5-Qt5                   5.15.2
PyQt5-sip                   12.12.1
pyqt5-tools                 5.15.9.3.3
pyreadline                  2.1
pyserial                    3.5
PySide6                     6.5.2
PySide6-Addons              6.5.2
PySide6-Essentials          6.5.2
PySocks                     1.7.1
python-dateutil             2.8.2
python-dotenv               0.21.1
pytorch-lightning           1.9.5
pytz                        2022.5
PyWavelets                  1.3.0
pywin32-ctypes              0.2.0
PyYAML                      6.0
pyzbar                      0.1.9
qt-material                 2.14
qt5-applications            5.15.2.2.3
qt5-tools                   5.15.2.1.3
QtPy                        2.3.1
qudida                      0.0.4
reportlab                   4.0.4
requests                    2.28.1
requests-oauthlib           1.3.1
rsa                         4.9
scikit-image                0.19.3
scikit-learn                1.0.2
scipy                       1.7.3
seaborn                     0.12.2
segmentation-models-pytorch 0.3.2
selenium                    4.11.2
sentry-sdk                  1.22.2
setuptools                  67.6.0
shapely                     2.0.1
shiboken6                   6.5.2
six                         1.16.0
sklearn                     0.0.post2
smmap                       5.0.0
sniffio                     1.3.0
sortedcontainers            2.4.0
sympy                       1.10.1
tap                         0.2
tensorboard                 2.10.0
tensorboard-data-server     0.6.1
tensorboard-plugin-wit      1.8.1
tensorboardX                2.6
termcolor                   2.3.0
thop                        0.1.1.post2209072238
threadpoolctl               3.1.0
tifffile                    2021.11.2
timm                        0.6.12
torch                       1.12.1+cu113
torchaudio                  0.12.1+cu113
torchgeometry               0.1.2
torchinfo                   1.7.2
torchmetrics                0.10.3
torchsummary                1.5.1
torchvision                 0.13.1+cu113
tqdm                        4.64.1
traitlets                   5.9.0
trio                        0.22.2
trio-websocket              0.10.3
typeshed-client             2.3.0
typing_extensions           4.3.0
ultralytics                 8.0.96
urllib3                     1.26.12
watchdog                    3.0.0
wcwidth                     0.2.6
Werkzeug                    2.2.2
wheel                       0.37.1
wsproto                     1.2.0
xlrd                        2.0.1
xlutils                     2.0.0
xlwt                        1.3.0
yarl                        1.9.2
zipp                        3.8.1
zxing                       1.0

A temporary password is generated for root@localhost: dLbGzsjQd7<O